+++ app/javascript/dashboard/i18n/locale/en/conversation.json
@@ -298,11 +298,24 @@
       "MACROS": "Macros",
       "SHOPIFY_ORDERS": "Shopify Orders"
     },
-
     "SHOPIFY": {
       "ORDER_ID": "Order #{id}",
       "ERROR": "Error loading Shopify orders",
-      "NO_SHOPIFY_ORDERS": "No Shopify orders found"
+      "NO_SHOPIFY_ORDERS": "No Shopify orders found",
+      "FINANCIAL_STATUS": {
+        "PENDING": "Pending",
+        "AUTHORIZED": "Authorized",
+        "PARTIALLY_PAID": "Partially Paid",
+        "PAID": "Paid",
+        "PARTIALLY_REFUNDED": "Partially Refunded",
+        "REFUNDED": "Refunded",
+        "VOIDED": "Voided"
+      },
+      "FULFILLMENT_STATUS": {
+        "FULFILLED": "Fulfilled",
+        "PARTIALLY_FULFILLED": "Partially Fulfilled",
+        "UNFULFILLED": "Unfulfilled"
+      }
     }
   },
   "CONVERSATION_CUSTOM_ATTRIBUTES": {
